Reorganize the headers and includes. The header include each others from relative path, and the others include headers from absolute path (Specified by option -I<DIR> from compiler).
